+++ src/main/java/kr/KWGraduate/BookPharmacy/global/infra/redis/refreshtoken/RefreshTokenRepository.java
@@ -1,9 +1,7 @@
-package kr.KWGraduate.BookPharmacy.global.infra.redis;
+package kr.KWGraduate.BookPharmacy.global.infra.redis.refreshtoken;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken,String> {
