+++ src/main/java/kr/KWGraduate/BookPharmacy/global/security/common/filter/JWTFilter.java
@@ -7,11 +7,11 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import kr.KWGraduate.BookPharmacy.global.security.common.dto.TokenDto;
-import kr.KWGraduate.BookPharmacy.global.infra.redis.RefreshToken;
+import kr.KWGraduate.BookPharmacy.global.infra.redis.refreshtoken.RefreshToken;
 import kr.KWGraduate.BookPharmacy.global.security.common.util.CookieType;
 import kr.KWGraduate.BookPharmacy.global.security.common.util.JWTUtil;
 import kr.KWGraduate.BookPharmacy.global.security.auth.service.ClientDetailsService;
-import kr.KWGraduate.BookPharmacy.global.infra.redis.RefreshTokenService;
+import kr.KWGraduate.BookPharmacy.global.infra.redis.refreshtoken.RefreshTokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.security.core.Authentication;
