+++ src/main/java/kr/KWGraduate/BookPharmacy/global/security/common/util/JWTUtil.java
@@ -7,9 +7,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import kr.KWGraduate.BookPharmacy.global.security.oauth.dto.CustomOauth2Client;
 import kr.KWGraduate.BookPharmacy.global.security.oauth.dto.Oauth2ClientDto;
 import kr.KWGraduate.BookPharmacy.global.security.common.dto.TokenDto;
-import kr.KWGraduate.BookPharmacy.global.infra.redis.RefreshToken;
+import kr.KWGraduate.BookPharmacy.global.infra.redis.refreshtoken.RefreshToken;
 import kr.KWGraduate.BookPharmacy.global.security.auth.service.ClientDetailsService;
-import kr.KWGraduate.BookPharmacy.global.infra.redis.RefreshTokenService;
+import kr.KWGraduate.BookPharmacy.global.infra.redis.refreshtoken.RefreshTokenService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
